+++ http-cache-quickcache/src/test.rs
@@ -125,6 +125,7 @@ async fn default_mode_with_options() -> Result<()> {
                 response_cache_mode_fn: None,
                 cache_bust: None,
                 cache_status_headers: true,
+                max_ttl: None,
             },
         }))
         .build();